requirements.txt
python-jose
bcrypt==4.1.3
passlib==1.7.4
pydantic-settings
fastapi
uvicorn
python-multipart
sqlalchemy